Initial implementation of Samza REST service.
This service allows clients to GET the status of all installed job instances or a particular instance.
It also allows clients to start/stop a job instance by running a PUT with form parameter "status" set to "started" or "stopped"

The primary abstraction is the JobProxy interface, which can be implemented or subclassed by users to customize details about how jobs are installed on the RM, how they are started, stopped, stat'ed, etc.

Future commits will address the following issues:
* Performance, by caching InstallationRecords and using the RM REST API rather than the CLI
* OSS markdown documentation