Reduce the time collision buffer to fix the ACG issue
After we change conainter timer metrics (such as choose_ns) from millisecond to nanosecond, it causes the timer metric to update a lot more often. As a result, it keeps a much larger amount of metric value inside its 5 minute storage. the default collision buffer is too large for a job such as ACG with has multiple thousand qps per container. For each timer metric, it adds about 100M memory usage, which causes constant GC in ACG job, and results in processing delay and fragmented call graphs. The fix here is to reduce the default buffer and have it configurable in the constructor.

RB=625896
G=samza-reviewers
R=yipan,jvenkatr
A=yipan