API Changes to make Standby Container Allocation Rack Aware (#1453)
Feature: The aim of this feature is to make all standby container requests rack aware such that all active containers and their corresponding standby containers are always on different racks. This helps with decreased downtime of applications during rack failures. One of the requirements of this feature is that the value of job.standbytasks.replication.factor is at max 2 for the rack awareness functionality to be honored.

Changes:

Added a new interface called FaultDomainManager which implements the YarnFaultDomainManager class for Yarn. This class takes care of the retrieval of node to rack information from Yarn.
Also defined what FaultDomain and FaultDomainType means.
Added config to enable making fault domain aware requests.
Added metrics to track fault domain aware requests.

API Changes: Added a new FaultDomainManager interface.